pyproject.toml
[project]
name = "doc-shape-shifter"
version = "0.1.0"
description = "Intelligent document format converter with automatic backend selection"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [{name = "Jon Chun", email = "jonchun2000@gmail.com"}]
readme = "README.md"
keywords = ["document", "conversion", "pdf", "markdown", "docx", "pandoc", "ocr", "rag"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Text Processing :: General",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Operating System :: OS Independent",
]

dependencies = [
    "click>=8.1",
    "python-magic>=0.4.27",
    "pymupdf>=1.24",
    "rich>=13.0",
]

[project.urls]
Homepage = "https://github.com/jon-chun/doc-shape-shifter"
Repository = "https://github.com/jon-chun/doc-shape-shifter"
Documentation = "https://github.com/jon-chun/doc-shape-shifter/blob/main/docs/user-manual.md"
Issues = "https://github.com/jon-chun/doc-shape-shifter/issues"
Changelog = "https://github.com/jon-chun/doc-shape-shifter/blob/main/CHANGELOG.md"

[project.optional-dependencies]
pandoc = ["pypandoc>=1.13"]
docling = ["docling>=2.82.0"]
markitdown = ["markitdown[all]>=0.1.5"]
tabula = ["tabula-py>=2.9"]
mathpix = ["requests>=2.31"]
ocr = ["surya-ocr>=0.6", "marker-pdf>=0.3.3"]
image = ["Pillow>=10.0", "opencv-python>=4.9"]
all = [
    "doc-shape-shifter[pandoc,docling,markitdown,tabula,mathpix]",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
    "mypy>=1.10",
]

[project.scripts]
dss = "doc_shape_shifter.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/doc_shape_shifter"]

[tool.ruff]
target-version = "py310"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
